refactor(pr-section): move tests out of src/cli into src/test
